What's New?
Aida can now effectively search through vairous url, and provide error message if data is not found, when data is not found, this mean the webscraping structure for the url has not be developed yet. In future i hope i can intergrate and AI ability in Aida that can make her auto update her code herself. But in meantime, i have to do that mannually.
Aida can also sort the data acquired from the webpage and next feature to expect is to get the top 5 companies from the data Aida has.